+++ libraries/pbes/test/eqelm_test.cpp
@@ -65,18 +65,18 @@ void test_pbes(const std::string& pbes_spec,
   my_pbes_rewriter pbesr(datar);
 
   // constelm algorithm
-  unsigned int log_level = 2;
+  size_t log_level = 2;
   pbes_eqelm_algorithm<pbes_expression, data::rewriter, my_pbes_rewriter> algorithm(datar, pbesr, log_level);
 
   // run the algorithm
-  algorithm.run(q); 
+  algorithm.run(q);
   BOOST_CHECK(q.is_well_typed());
   if (!q.is_well_typed())
   {
     std::cerr << pp(pbes_to_aterm(q)) << std::endl;
   }
 
-  pbes_system::detail::pbes_property_map info1(q);  
+  pbes_system::detail::pbes_property_map info1(q);
   pbes_system::detail::pbes_property_map info2(expected_result);
   std::string diff = info1.compare(info2);
   if (!diff.empty())
