+++ tools/diagraphica/corrlplot.h
@@ -22,44 +22,44 @@
 
 class CorrlPlot : public Visualizer
 {
-public:
+  public:
     // -- constructors and destructor -------------------------------
     CorrlPlot(
-        Mediator* m,
-        Graph* g,
-        GLCanvas* c );
+      Mediator* m,
+      Graph* g,
+      GLCanvas* c);
     virtual ~CorrlPlot();
 
     // -- set data functions ----------------------------------------
     void setValues(
-        const int &idx1,
-        const int &idx2,
-        const std::vector< std::vector< int > > &mapXY,
-        const std::vector< std::vector< int > > &num );
+      const size_t& idx1,
+      const size_t& idx2,
+      const std::vector< std::vector< size_t > > &mapXY,
+      const std::vector< std::vector< int > > &num);
     void clearValues();
 
-  	void setDiagram( Diagram* dgrm );
+    void setDiagram(Diagram* dgrm);
 
     // -- visualization functions  ----------------------------------
-    void visualize( const bool &inSelectMode );
+    void visualize(const bool& inSelectMode);
     void drawAxes(
-        const bool &inSelectMode,
-        const std::string &xLbl,
-        const std::string &yLbl );
-    void drawLabels( const bool &inSelectMode );
-    void drawPlot( const bool &inSelectMode );
-    void drawDiagram( const bool &inSelectMode );
+      const bool& inSelectMode,
+      const std::string& xLbl,
+      const std::string& yLbl);
+    void drawLabels(const bool& inSelectMode);
+    void drawPlot(const bool& inSelectMode);
+    void drawDiagram(const bool& inSelectMode);
 
     // -- input event handlers --------------------------------------
     void handleMouseMotionEvent(
-		const int &x,
-		const int &y );
+      const int& x,
+      const int& y);
     /*
     void handleMouseEnterEvent();
     void handleMouseLeaveEvent();
     */
 
-protected:
+  protected:
     // -- utility data functions ------------------------------------
     void initLabels();
     void calcMaxNumber();
@@ -69,25 +69,25 @@ protected:
     //void clear();
     void setScalingTransf();
     void displTooltip(
-        const int &xIdx,
-        const int &yIdx );
+      const int& xIdx,
+      const int& yIdx);
 
     void calcPositions();
     void clearPositions();
 
     // -- hit detection ---------------------------------------------
     void processHits(
-        GLint hits,
-        GLuint buffer[] );
+      GLint hits,
+      GLuint buffer[]);
 
     // -- data members ----------------------------------------------
 
     // data
-    int attrIdx1;
-    int attrIdx2;
+    size_t attrIdx1;
+    size_t attrIdx2;
     std::string xLabel;
     std::string yLabel;
-    std::vector< std::vector< int > > mapXToY; // for each x idx, 1 or more y indices
+    std::vector< std::vector< size_t > > mapXToY; // for each x idx, 1 or more y indices
     std::vector< std::vector< int > > number;  // number associated with every entry in mapXToY
     std::vector< int > maxNumX;   // for every column, the largest number
     int sumMaxNumX;          // sum of entries in maxNumX
@@ -106,8 +106,8 @@ protected:
     double     scaleDgrm;       // scale factor for diagram
     Position2D posDgrm;         // positions of diagram
     bool       showDgrm;        // show or hide diagram
-    int        attrValIdx1Dgrm; // value idx of 1st attribute associated with diagram
-    int        attrValIdx2Dgrm; // value idx of 2nd attribute associated with diagram
+    size_t        attrValIdx1Dgrm; // value idx of 1st attribute associated with diagram
+    size_t        attrValIdx2Dgrm; // value idx of 2nd attribute associated with diagram
     std::string     msgDgrm;        // message to show with diagram
 };
