+++ tools/diagraphica/edge.h
@@ -19,20 +19,20 @@ class Node;
 
 class Edge
 {
-public:
-	// -- constructors and destructor -------------------------------
+  public:
+    // -- constructors and destructor -------------------------------
     Edge(
-        const int &idx,
-        const std::string &lbl );
-	virtual ~Edge();
+      const size_t& idx,
+      const std::string& lbl);
+    virtual ~Edge();
 
     // -- set functions ---------------------------------------------
-    void setInNode( Node* in );
-    void setOutNode( Node* out );
-    void setBundle( Bundle* b );
+    void setInNode(Node* in);
+    void setOutNode(Node* out);
+    void setBundle(Bundle* b);
 
     // -- get functions ---------------------------------------------
-    int getIndex();
+    size_t getIndex();
     std::string getLabel();
     Node* getInNode();
     Node* getOutNode();
@@ -43,12 +43,12 @@ public:
     void clearOutNode();
     void clearBundle();
 
-protected:
+  protected:
     // -- data members ----------------------------------------------
-	int index;      // index in list of graph edges
+    size_t index;      // index in list of graph edges
     std::string label;   // "action label"
     Node* inNode;   // association
-	Node* outNode;  // association
+    Node* outNode;  // association
     Bundle* bundle; // association
 };
